# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Job Vacancy API
  version: 0.6.0
  description: |
    Shared contract for the web (React) and mobile (Flutter) clients.

    **Realtime.** POST /ws/ticket, then open `wss://…/api/v1/ws?ticket=…`. Server events:
    `ready`, `message.new`, `message.read`, `message.deleted`, `typing`, `presence`,
    `presence.snapshot`, `notification.new`, `pong`, `error` — each `{"type", "data"}`.
    Client frames: `{"type":"ping"}`, `{"type":"typing","conversation_id"}`,
    `{"type":"read","conversation_id","message_id"}`, `{"type":"watch_presence","user_ids":[…]}`.
    Events are not buffered: after reconnecting, fetch `/conversations/{id}/messages?after=<last id>`.

    **Uploads.** POST /files → multipart POST to `upload.url` with every `upload.fields`
    entry followed by the `file` field → POST /files/{id}/complete → use the file id.

    **Envelope.** Success: `{"data": ...}` (lists add `"meta"`). Failure:
    `{"error": {"code", "message", "fields"?, "retry_after"?}}`. `code` is stable; clients
    translate it into uz / uz-Cyrl / ru / en. `fields` maps a JSON field to the failing
    rule (`required`, `email`, `min=8`, `password`, `oneof=...`).

    **Deadlines.** Every request runs under a server deadline (8 s; PDF export 25 s;
    the WebSocket is exempt) and each database statement under 5 s. When one is exceeded
    the API answers `503` with code `timeout` and `Retry-After: 1`; the call is safe to
...
    limits of their own.

...

...
    it rendered.

...

...
servers:
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    Cursor: { name: cursor, in: query, schema: { type: string } }
...
  schemas:
...
    Error:
      type: object
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string, examples: [invalid_credentials] }
            message: { type: string }
            fields: { type: object, additionalProperties: { type: string } }
            retry_after: { type: integer, description: seconds }
...
    AdminUser:
      type: object
      required: [id, email, phone, full_name, avatar_url, role, status, email_verified, phone_verified, last_seen_at, created_at, deleted_at]
      properties:
        id: { type: string, format: uuid }
        email: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        full_name: { type: string, description: "empty for deleted accounts" }
        avatar_url: { type: [string, "null"] }
        role: { type: string, enum: [seeker, employer, admin] }
        status: { type: string, enum: [active, blocked, deleted] }
        email_verified: { type: boolean }
        phone_verified: { type: boolean }
        last_seen_at: { type: [string, "null"], format: date-time }
        created_at: { type: string, format: date-time }
        deleted_at: { type: [string, "null"], format: date-time }
    AdminUserDetail:
      allOf:
        - $ref: "#/components/schemas/AdminUser"
        - type: object
          required: [companies]
          properties:
            companies:
              type: array
              items:
                type: object
                required: [id, name, slug, status, role]
                properties:
                  id: { type: string, format: uuid }
                  name: { type: string }
                  slug: { type: string }
                  status: { type: string, enum: [active, blocked, deleted] }
                  role: { type: string, enum: [owner, admin, recruiter] }
...
    AdminSkill:
      type: object
      required: [id, name, slug, is_verified, usage_count, vacancies, resumes, created_at]
      properties:
        id: { type: integer }
        name: { type: string }
        slug: { type: string }
        is_verified: { type: boolean }
        usage_count: { type: integer }
        vacancies: { type: integer }
        resumes: { type: integer }
        created_at: { type: string, format: date-time }
    AdminStats:
      type: object
      required: [totals, days]
      properties:
        totals:
          type: object
          required: [users, seekers, employers, blocked_users, companies, published_vacancies, moderation_queue, open_reports]
          properties:
            users: { type: integer }
            seekers: { type: integer }
            employers: { type: integer }
            blocked_users: { type: integer }
            companies: { type: integer }
            published_vacancies: { type: integer }
            moderation_queue: { type: integer }
            open_reports: { type: integer }
        days:
          type: array
          description: oldest first, Asia/Tashkent days
          items:
            type: object
            required: [date, registrations, vacancies, applications]
            properties:
              date: { type: string, format: date }
              registrations: { type: integer }
              vacancies: { type: integer }
              applications: { type: integer }
    AuditEntry:
      type: object
      required: [id, action, object_type, object_id, details, ip, created_at, admin]
      properties:
        id: { type: string, format: uuid }
        action: { type: string, examples: [user.block, company.block, vacancy.approve, skill.merge, report.resolved] }
        object_type: { type: string, enum: [user, company, vacancy, skill, report, search_term] }
        object_id: { type: string }
        details: { type: object }
        ip: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        admin:
          type: [object, "null"]
          properties: { id: { type: string, format: uuid }, full_name: { type: string }, email: { type: [string, "null"] } }
...
  responses:
...
    Error:
      description: Error
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
paths:
...
  /admin/users:
    get:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Search users by part of name, e-mail or phone; newest first
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string, enum: [seeker, employer, admin] } }
        - { name: status, in: query, schema: { type: string, enum: [active, blocked, deleted] } }
        - { $ref: "#/components/parameters/Cursor" }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Page
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/AdminUser" } }, meta: { type: object, properties: { next_cursor: { type: [string, "null"] } } } } } } }
  /admin/users/{user}:
    get:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: One user with their company memberships
      parameters: [{ name: user, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: User
          content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/AdminUserDetail" } } } } }
        "404": { $ref: "#/components/responses/Error" }
  /admin/users/{user}/block:
    post:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Block a user — sign-in refused, all sessions revoked at once, their published/pending vacancies unpublished
      parameters: [{ name: user, in: path, required: true, schema: { type: string, format: uuid } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [reason], properties: { reason: { type: string, minLength: 3, maxLength: 500 } } } } }
      responses:
        "200":
          description: User
          content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/AdminUser" } } } } }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /admin/users/{user}/unblock:
    post:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Unblock a user
      parameters: [{ name: user, in: path, required: true, schema: { type: string, format: uuid } }]
      responses:
        "200":
          description: User
          content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/AdminUser" } } } } }
        "409": { $ref: "#/components/responses/Error" }
  /admin/skills:
    get:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Skills newest first with usage counts (verified=false is the moderation queue)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: verified, in: query, schema: { type: boolean } }
        - { $ref: "#/components/parameters/Cursor" }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Page
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/AdminSkill" } }, meta: { type: object, properties: { next_cursor: { type: [string, "null"] } } } } } } }
  /admin/skills/{skill}/verification:
    parameters: [{ name: skill, in: path, required: true, schema: { type: integer } }]
    put:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Mark a skill verified (suggested to everyone)
      responses:
        "200":
          description: Skill
          content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/AdminSkill" } } } } }
        "404": { $ref: "#/components/responses/Error" }
    delete:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Remove the verified mark
      responses:
        "200":
          description: Skill
          content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/AdminSkill" } } } } }
        "404": { $ref: "#/components/responses/Error" }
  /admin/skills/{skill}/merge:
    post:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Merge a duplicate skill into another; vacancies and resumes move over and the duplicate is deleted
      parameters: [{ name: skill, in: path, required: true, schema: { type: integer } }]
      requestBody:
        required: true
        content: { application/json: { schema: { type: object, required: [into_id], properties: { into_id: { type: integer } } } } }
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    required: [into, vacancies, resumes]
                    properties: { into: { $ref: "#/components/schemas/AdminSkill" }, vacancies: { type: integer }, resumes: { type: integer } }
        "400": { $ref: "#/components/responses/Error" }
        "404": { $ref: "#/components/responses/Error" }
...
  /admin/stats:
    get:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Dashboard totals and daily sign-ups, vacancies and applications
      parameters: [{ name: days, in: query, schema: { type: integer, minimum: 1, maximum: 90, default: 30 } }]
      responses:
        "200":
          description: Stats
          content: { application/json: { schema: { type: object, properties: { data: { $ref: "#/components/schemas/AdminStats" } } } } }
  /admin/audit-log:
    get:
      tags: [admin]
      security: [{ bearer: [] }]
      summary: Admin actions, newest first
      parameters:
        - { name: object_type, in: query, schema: { type: string } }
        - { name: object_id, in: query, schema: { type: string } }
        - { name: admin_id, in: query, schema: { type: string, format: uuid } }
        - { $ref: "#/components/parameters/Cursor" }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 20 } }
      responses:
        "200":
          description: Page
          content: { application/json: { schema: { type: object, properties: { data: { type: array, items: { $ref: "#/components/schemas/AuditEntry" } }, meta: { type: object, properties: { next_cursor: { type: [string, "null"] } } } } } } }
...